+++ src/shift_left/tests/data/flink-project/pipelines/dimensions/c360/dim_groups/sql-scripts/dml.c360_dim_groups.sql
@@ -1,13 +1,14 @@
 INSERT INTO sl_c360_dim_groups
 SELECT
   group_id,
-  tenant_id,
+  t.tenant_id,
   group_name,
   group_type,
   tenant_name,
   created_date,
   is_active,
   updated_at
-FROM sl_c360_src_groups g
-join sl_cmn_src_tenants tenant
-on g.tenant_id = sl_tenant.tenant_id
+FROM sl_c360_src_groups as g
+JOIN sl_cmn_src_tenants as t
+    ON g.tenant_id = t.tenant_id
+WHERE t.tenant_id IS NOT NULL;