+++ src/shift_left/tests/data/flink-project/pipelines/facts/c360/fct_user_per_group/sql-scripts/ddl.c360_fct_user_per_group.sql
@@ -1,14 +1,12 @@
 CREATE TABLE IF NOT EXISTS sl_c360_fct_user_per_group (
+  tenant_id STRING NOT NULL,
   group_id STRING NOT NULL,
   group_name STRING,
-  group_type STRING,
   total_users BIGINT,
   active_users BIGINT,
   inactive_users BIGINT,
-  latest_user_created_date BIGINT,
-  fact_updated_at TIMESTAMP,
-  PRIMARY KEY(group_id) NOT ENFORCED
-) DISTRIBUTED BY HASH(group_id) INTO 1 BUCKETS
+  PRIMARY KEY(tenant_id,group_id) NOT ENFORCED
+) DISTRIBUTED BY HASH(tenant_id,group_id) INTO 1 BUCKETS
 WITH (
   'changelog.mode' = 'upsert',
   'key.format' = 'avro-registry',