+++ src/shift_left/tests/data/flink-project/pipelines/sources/c360/src_users/sql-scripts/dml.src_c360_users.sql
@@ -30,7 +30,6 @@ WITH deduplicated_users AS (
   WHERE
     user_id IS NOT NULL  -- Ensure we have valid user_id
     AND user_email IS NOT NULL  -- Ensure we have valid email
-    AND is_active
 )
 SELECT
   user_id,