+++ src/shift_left/tests/data/flink-project/pipelines/sources/common/src_tenant/sql-scripts/ddl.src_common_tenant.sql
@@ -8,7 +8,7 @@ CREATE TABLE IF NOT EXISTS sl_cmn_src_tenants (
   PRIMARY KEY(tenant_id) NOT ENFORCED
 ) DISTRIBUTED BY HASH(tenant_id) INTO 1 BUCKETS
 WITH (
-  'changelog.mode' = 'append',
+  'changelog.mode' = 'upsert',
   'key.format' = 'avro-registry',
   'value.format' = 'avro-registry',
   'kafka.retention.time' = '0',