+++ src/shift_left/tests/data/flink-project/pipelines/test_prepare_tables_integration.sql
@@ -2,12 +2,44 @@
 -- This file contains safe table preparation statements for testing
 
 -- Comment line that should be skipped
-ALTER TABLE `src_table_1` SET ('connector' = 'kafka');
+ALTER TABLE `orphan` SET ('changelog.mode' = 'upsert');
 
 -- Another comment with some details
-ALTER TABLE `src_table_1` SET ('topic' = 'test-topic-1');
+alter table orphan SET ('scan.startup.mode' = 'latest-offset')
 
 -- Set format properties
-ALTER TABLE `src_table_1` SET ('value.format' = 'json');
+ALTER TABLE `orphan` SET ('value.format' = 'json-registry');
 
--- This is a comment that should be ignored
+ALTER TABLE orphan SET ('kafka.cleanup-policy' = 'compact');
+
+-- Supported options:
+--
+-- changelog.mode
+-- connector
+-- error-handling.log.target
+-- error-handling.mode
+-- kafka.cleanup-policy
+-- kafka.compaction.time
+-- kafka.consumer.isolation-level
+-- kafka.max-message-size
+-- kafka.message-timestamp-type
+-- kafka.producer.compression.type
+-- kafka.retention.size
+-- kafka.retention.time
+-- key.avro-registry.id-encoding
+-- key.avro-registry.schema-context
+-- key.avro-registry.subject-names
+-- key.fields-prefix
+-- key.format
+-- late-handling.mode
+-- scan.bounded.mode
+-- scan.bounded.specific-offsets
+-- scan.bounded.timestamp-millis
+-- scan.startup.mode
+-- scan.startup.specific-offsets
+-- scan.startup.timestamp-millis
+-- value.avro-registry.id-encoding
+-- value.avro-registry.schema-context
+-- value.avro-registry.subject-names
+-- value.fields-include
+-- value.format